fix(engine): renaming the test
